improve session permission handling
