+++ src/types/next-auth-helper.ts
@@ -1,0 +1,5 @@
+export enum QboPermission {
+  Accounting = 0,
+  Profile,
+  None,
+}
